completed controllers index.js, and moved the seeds files into the models folders. also added a retailers.json page for seed data
